fix(test): move process-environment stub under terminal harness
Keep tests/unit/main/terminal at the directory density hard cap.
