+++ tests/unit/main/terminal/debug.test.ts
@@ -131,7 +131,7 @@ describe("terminal native debug IPC", () => {
     }));
 
     const { stubTerminalIpcProcessEnvironment } = await import(
-      "./stub-process-environment.ts"
+      "./harness/stub-process-environment.ts"
     );
     const { registerTerminalIpc } = await import("@main/ipc/terminal/index.ts");
     registerTerminalIpc(fakeIpcMain as never, {
