+++ tests/unit/main/terminal/harness/stub-process-environment.ts
