+++ tests/unit/main/terminal/panel-id-scoping.test.ts
@@ -140,7 +140,7 @@ describe("multi-window panel id scoping (#16 #30)", () => {
 
     const closeTerminal = fakeAddon.closeTerminal;
     const { stubTerminalIpcProcessEnvironment } = await import(
-      "./stub-process-environment.ts"
+      "./harness/stub-process-environment.ts"
     );
     const { registerTerminalIpc } = await import("@main/ipc/terminal/index.ts");
     registerTerminalIpc(fakeIpcMain as never, {
@@ -266,7 +266,7 @@ describe("multi-window panel id scoping (#16 #30)", () => {
     };
 
     const { stubTerminalIpcProcessEnvironment } = await import(
-      "./stub-process-environment.ts"
+      "./harness/stub-process-environment.ts"
     );
     const { registerTerminalIpc } = await import("@main/ipc/terminal/index.ts");
     registerTerminalIpc(fakeIpcMain as never, {
